+++ src/lib/offchain-lib/TxBuilding/Transactions.hs
@@ -179,9 +179,6 @@ mintOracleNFTAndLockDatum providerCtx skey adminPKH = do
               opMinOutputLovelace = 3500000
             }
 
-    -- Oracle validator address
-    oracleAddr <- scriptAddress oracleValidatorGY
-
     -- Spend seed UTxO
     let spendSeed = mustHaveInput (GYTxIn seedGYRef GYTxInWitnessKey)
 
@@ -191,14 +188,7 @@ mintOracleNFTAndLockDatum providerCtx skey adminPKH = do
     let mintNFT = mustMint mp gyRedeemer oracleNFTTN 1
 
     -- Lock oracle NFT + datum at oracle validator address
-    let lockOutput =
-          mustHaveOutput
-            GYTxOut
-              { gyTxOutAddress = oracleAddr,
-                gyTxOutDatum = Just (datumFromPlutusData initialOracleParams, GYTxOutUseInlineDatum),
-                gyTxOutValue = valueSingleton theOracleNFTAC 1 <> valueFromLovelace 3500000,
-                gyTxOutRefS = Nothing
-              }
+    lockOutput <- txMustLockStateWithInlineDatumAndValue oracleValidatorGY initialOracleParams (valueSingleton theOracleNFTAC 1 <> valueFromLovelace 3500000)
 
     body <- buildTxBody $ mconcat [spendSeed, mintNFT, lockOutput]
     return (body, theOracleNFTAC)